Add `ChainlinkStatusEntry` interface and improve scope chain parsing.

+++ idl-clients-kamino/src/main/java/software/sava/idl/clients/kamino/scope/entries/FixedPrice.java
@@ -1,9 +1,8 @@
 package software.sava.idl.clients.kamino.scope.entries;
 
 import software.sava.idl.clients.kamino.scope.gen.types.OracleType;
-import software.sava.idl.clients.kamino.scope.gen.types.Price;
 
-public record FixedPrice(long value, long exp) implements ScopeEntry {
+public record FixedPrice(long value, int exp) implements ScopeEntry {
 
   @Override
   public OracleType oracleType() {
